implement student interface with dummy data of student entity and CRUD operation

+++ university.catalog/Entities/Student.Entity/StudentEntity.cs
@@ -12,11 +12,11 @@ namespace University.Catalog.Entities
         public DateTimeOffset StudentEnrollmentDate { get; init; }
         public string? CountryName { get; init; }
         public string? StudentGender { get; init; }
-        public int StudentContactNumber { get; init; }
+        public long StudentContactNumber { get; init; }
         public bool IsScholar { get; init; }
         public string? CoreBranchName { get; init; }
         public string? EmergencyContactName { get; init; }
-        public int EmergencyContactNumber { get; init; }
+        public long EmergencyContactNumber { get; init; }
         public string? BloodGroup { get; init; }
         public bool IsDayScholar { get; init; }
         public int DurationForCourse { get; init; }
